requirements.txt
streamlit==1.36.0
langchain==0.2.7
langchain-community==0.2.6
pypdf==4.2.0
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.2.2
faiss-cpu==1.8.0
sentence-transformers==2.7.0
tiktoken==0.7.0
# openai==1.35.1 # Commented out as we are no longer using OpenAI
transformers==4.42.3
torch==2.3.1 # Required by transformers and sentence-transformers
requests # <--- ADDED THIS FOR GEMINI API CALLS
